requirements.txt
pandas
pathlib
tqdm
matplotlib
pandas_datareader
seaborn
statsmodels
scikit-learn
# TA-Lib
futu-api
clickhouse-driver
confluent-kafka
python-dotenv
mysql-connector-python
# git+https://github.com/happydasch/btplotting